kms_content_protection: srm and downstream_info test
Retrieve the downstream info and use that to stitch a srm and
verify the revocation process.

Adds a connector properties called CP_downstream_info

SRM stitched here is polluted. Kernel trusts the userspace to validate
the SRM integrity through DCP signature. Actually this test exploits
that trust to test SRM and downstream_info features.

HDCP1.4 BKSV and HDCP2.2 receiver id of a HDCP sink is different. So we
need to force the KMD to use a single HDCP version across
downstream_info gathering and SRM setting and verifying teh revocation
process. So Type 1 is used.

v2:
  binary sysfs is used to write the srm into kernel [Daniel]

Signed-off-by: Ramalingam C <ramalingam.c@intel.com>